java_testgen_app/src/test/java/Copilot/DepthFirstSearch/PromptOnly/T7/DepthFirstSearchTest.java
package Copilot.DepthFirstSearch.PromptOnly.T7;

import static org.junit.Assert.assertArrayEquals;

import java.util.Arrays;
import java.util.List;

import org.junit.Test;

public class DepthFirstSearchTest {
    @Test
    public void test1() {
        List<DepthFirstSearch.Edge> edges = Arrays.asList(
                new DepthFirstSearch.Edge(0, 1),
                new DepthFirstSearch.Edge(0, 3),
                new DepthFirstSearch.Edge(1, 4),
                new DepthFirstSearch.Edge(2, 4),
                new DepthFirstSearch.Edge(2, 5),
                new DepthFirstSearch.Edge(3, 1),
                new DepthFirstSearch.Edge(4, 3),
                new DepthFirstSearch.Edge(5, 5),
                new DepthFirstSearch.Edge(6, 6));
        DepthFirstSearch dfs = new DepthFirstSearch(edges, 7);
        dfs.dfs(0, new boolean[7]);
        assertArrayEquals(new Integer[] { 0, 1, 4, 3, 2, 5 }, dfs.visitedNodes.toArray());
    }

    @Test
    public void test2() {
        List<DepthFirstSearch.Edge> edges = Arrays.asList(
                new DepthFirstSearch.Edge(0, 1),
                new DepthFirstSearch.Edge(0, 3),
                new DepthFirstSearch.Edge(1, 4),
                new DepthFirstSearch.Edge(2, 4),
                new DepthFirstSearch.Edge(2, 5),
                new DepthFirstSearch.Edge(3, 1),
                new DepthFirstSearch.Edge(4, 3),
                new DepthFirstSearch.Edge(5, 5),
                new DepthFirstSearch.Edge(6, 6));
        DepthFirstSearch dfs = new DepthFirstSearch(edges, 7);
        dfs.dfs(2, new boolean[7]);
        assertArrayEquals(new Integer[] { 2, 4, 3, 1, 0, 5 }, dfs.visitedNodes.toArray());
    }

    @Test
    public void test3() {
        List<DepthFirstSearch.Edge> edges = Arrays.asList(
                new DepthFirstSearch.Edge(0, 1),
                new DepthFirstSearch.Edge(0, 3),
                new DepthFirstSearch.Edge(1, 4),
                new DepthFirstSearch.Edge(2, 4),
                new DepthFirstSearch.Edge(2, 5));
    }
}